Enable M4T vocoder inference on the GPU in fp16. (#151)
* Enable M4T vocoder inference on the GPU, remove unnecessary device to host syncs in the Translator.

* Address comment, disable non-deterministic GPU integ test.

+++ tests/integration/models/test_pretssel_vocoder.py
@@ -27,7 +27,7 @@ def test_pretssel_vocoder(example_rate16k_audio: AudioDecoderOutput) -> None:
 
     feat = convert_to_collated_fbank(audio_dict, dtype=dtype)["seqs"][0]
 
-    vocoder = load_mel_vocoder_model("vocoder_mel", device=device, dtype=torch.float32)
+    vocoder = load_mel_vocoder_model("vocoder_mel", device=device, dtype=dtype)
     vocoder.eval()
 
     with torch.inference_mode():
